Install franz only if not already installed

+++ src/utility_franz.sh
@@ -3,14 +3,18 @@
 source "$(dirname "$0")/_base.sh"
 
 my_step_begin "install franz"
-installer_path=/tmp/franz_installer.deb
-installer_url=$( \
-	curl -sS https://api.github.com/repos/meetfranz/franz/releases/latest \
-	| grep browser_download_url	 \
-	| grep .deb \
-	| sed -E 's/.*"([^"]+)".*/\1/' \
-)
-curl -L $installer_url -o $installer_path
-sudo dpkg --install $installer_path
-rm /tmp/franz_installer.deb
+if my_command_exists franz; then
+	echo 'franz already installed'
+else
+	installer_path=/tmp/franz_installer.deb
+	installer_url=$( \
+		curl -sS https://api.github.com/repos/meetfranz/franz/releases/latest \
+		| grep browser_download_url	 \
+		| grep .deb \
+		| sed -E 's/.*"([^"]+)".*/\1/' \
+	)
+	curl -L $installer_url -o $installer_path
+	sudo dpkg --install $installer_path
+	rm /tmp/franz_installer.deb
+fi
 my_step_end
